contrib(profiling): add JfrEventUtils - to expose getEventName
To determine an event name of a known event class, as it will be recorded.
Either it was explicitly set via @Name or it will be the full class name.
